moved sampling to data utils
