+++ data/data_utils.py
@@ -1,9 +1,14 @@
-import os
+import os, sys
+import glob
 import numpy as np
 import pandas as pd
 import random
 import gym
+import torch
+from torch.utils.data import WeightedRandomSampler
 
+sys.path.append(os.path.abspath(os.path.join('..')))
+from config import Config as cfg
 
 def rollout_viz(env, data, seed=None, render=False):
     """Visualize previously saved rollouts in the gym env"""
@@ -34,6 +39,49 @@ def rollout_viz(env, data, seed=None, render=False):
         steps += 1
         if done: break
     return total_reward
+
+
+def get_weighted_sampler(ratio, normalize=False):
+    """
+    Returns dataframe and WeightedRandomSampler based on desired ratio of sub-optimal to 
+    optimal data rollouts
+        ratio: split tuple of opt to sub-opt data, e.g. (0.1, 0.9) for
+               10% optimal and 90% suboptimal training data
+        normalize: whether the weights should be normalized based on count
+    """
+
+    assert np.sum(ratio) == 1.0, "Ratios should some up to 1.0"
+    files = glob.glob(os.path.join(cfg.DATA_PATH, "*.csv"))
+    weights = torch.Tensor()
+    df = pd.DataFrame()
+    rollout_mapping = {"none": 0, "laggy": 1, "noisy": 2}
+    rollout_count = {"opt": 0, "subopt": 0}
+    for file in files:
+        rollout = pd.read_csv(file, header=0, index_col=None)
+        rollout_type = file.split("_")[1]
+        # Append rollout type column to dataframe
+        rollout['type'] = rollout_mapping[rollout_type]
+
+        if rollout_type == cfg.OPT_LABEL:
+            rollout_count["opt"] += len(rollout)
+            weights = torch.cat((weights, ratio[0]*torch.ones(len(rollout))))
+        else:
+            rollout_count["subopt"] += len(rollout)
+            weights = torch.cat((weights, ratio[1]*torch.ones(len(rollout))))
+
+        df = pd.concat((df, rollout))
+
+    # Balancing based on opt vs subopt data availability
+    if normalize:
+        # N = len(weights)
+        weights[weights==ratio[0]] /= (rollout_count["opt"])#*ratio[0]
+        weights[weights==ratio[1]] /= (rollout_count["subopt"])#*ratio[1]
+        # weights[weights==ratio[0]] = (N/rollout_count["opt"]*ratio[0])
+        # weights[weights==ratio[1]] = (N/rollout_count["subopt"]*ratio[1])
+
+    # NOTE: sampeld with replacement by default
+    sampler = WeightedRandomSampler(weights, len(weights))
+    return df, sampler, weights
